requirements.txt
# Web Framework
flask==3.1.2
flask-cors==6.0.1

# MQTT Client
paho-mqtt==2.1.0

# Database
peewee==3.18.3

# Configuration
python-dotenv==1.2.1

# Utilities
python-dateutil==2.9.0.post0
pyserial==3.5
requests==2.32.5